Simplify references to Work
Aaron had thought we should "Probably use different language for the
'Work' performed by the project and the work performed by Members for
Projects." We did this in 01e8cb2a, but we didn't end up referring to
"Project Tasks" anywhere except where that was defined. I think by
switching to Collaborators we make it clear enough the these are people
collaborating in the Project's Work.